% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow_1Gene_CT_correlations.R
\name{slidingWindow_1Gene_CT_correlations}
\alias{slidingWindow_1Gene_CT_correlations}
\title{Identify all genes that correlate to a gene of interest in a cell type of interest}
\usage{
slidingWindow_1Gene_CT_correlations(
  GeneExp,
  CTproportions,
  GOI_anchor,
  CT_anchor,
  CT_cor,
  Type_pValue = "p.value",
  Significance_cutoff = 0.05,
  Correlation_cutoff = 0.5
)
}
\arguments{
\item{GeneExp}{numeric matrix with samples as columns and features as rows}

\item{CTproportions}{numeric matrix with samples as rows and cell type names as columns}

\item{GOI_anchor}{gene to order samples by}

\item{CT_anchor}{cell type of anchor gene}

\item{CT_cor}{cell type where you want to identify correlated genes}

\item{Type_pValue}{p.value cut off FDR or p.value}

\item{Significance_cutoff}{significance must be less than this to be reports}

\item{Correlation_cutoff}{correlation must be less than this to be reports}
}
\value{
reports p.value, FDR, and correlation for genes passing thresholds in a given cell type
}
\description{
This function takes bulk gene expression and esimated cell type proportions
and performs sliding window deconvolution. The output is all genes that correlate.
to a gene of interest in a specific cell type. It assumes that both input
matrixes are numeric with features (expression) or samples (proportion) as row names.
Columns of the expression matrix are samples and rows of the proportion matrix are
samples. Only samples present in both matrixes will be used. There must be more
than 40 samples total.
}
