% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow_2Gene_2CT.R
\name{slidingWindow_2Gene_2CT}
\alias{slidingWindow_2Gene_2CT}
\title{Correlate 2 genes in a cell type specific manner}
\usage{
slidingWindow_2Gene_2CT(
  GeneExp,
  CTproportions,
  GOI_anchor,
  GOI_cor,
  CT_anchor,
  CT_cor
)
}
\arguments{
\item{GeneExp}{numeric matrix with samples as columns and features as rows}

\item{CTproportions}{numeric matrix with samples as rows and cell type names as columns}

\item{GOI_anchor}{gene to order samples by}

\item{GOI_cor}{gene you want to correlate to}

\item{CT_anchor}{cell type of anchor gene}

\item{CT_cor}{cell type of gene you want to correlate to}
}
\value{
Correlation, p.value, and figures of bulk and cell type specific expression of genes
}
\description{
This function takes bulk gene expression and esimated cell type proportions
and performs sliding window deconvolution. The output is correlation of 2
features of interest in a cell type specific manner. It assumes that both input
matrixes are numeric with features (expression) or samples (proportion) as row names.
Columns of the expression matrix are samples and rows of the proportion matrix are
samples. Only samples present in both matrixes will be used. There must be more
than 40 samples total.
}
